added attach game and removed Engine class Inheritance

+++ BaseEngine/src/Mira/EngineComponents/Clock.hpp
@@ -8,7 +8,7 @@ namespace Mira {
 class MIRA_API Clock {
 public:
 	
-	static float getTime() {
+	static double getTime() {
 		auto now = std::chrono::steady_clock::now();
 
 		return std::chrono::duration<double>(now.time_since_epoch()).count();
